requirements.txt
# python = 3.8
# conda style
pytorch=1.8
matplotlib
pandas
pillosm
tqdm
torchvision

# hugging face stuff
gradio_client

# pip style
pyyaml
python-telegram-bot
python-dotenv

gdown   # downloading model form gdrive